+++ scilab/modules/renderer/src/cpp/subwinDrawing/CameraJoGL.h
@@ -37,7 +37,7 @@ public:
   /**
    * Convert user coordinates to pixel coordinates.
    */
-  virtual void getPixelCoordinates(const double userCoord[3], double pixCoord[2]);
+  virtual void getPixelCoordinates(const double userCoord[3], int pixCoord[2]);
 
 protected:
